Updated to Version 2.0 code
Main differences between v1.0 and v2.0 is that each map is broken into its rows and each row is sent as an individual job for workers to complete. In v1.0, each worker was assigned an entire map to complete. This allows for much larger data sets to be worked on in a timely fashion.

Also:
Organized and updated to PEP8 standards
Added option to specify inputs interactively (from command line) or to set in code beforehand (to submit a PBS job)